PP-590 Connector spamming logs with JDBC operations * All eclipselink configurations for JPA are moved to the config.yaml Our logs are been spammed by JDBC operations due to the eclipselink log configurations. The log configurations were in the persistence.xml. These configurations are now moved to the config.yaml and the log levels can now be toggled through environment variables. Unfortunately not all configurations are in the persistence.xml could be moved as eclipselink JPA needs it for JPA to work.
* Allowing adding defaults in the config.yml.
And with this change log level defaults are provided in the config.yaml.
For this to work, the instance creation of
EnvironmentVariableSubstitutor has been changed. A false flag is passed
into this, which makes it non-strict so that we can have defaults in the
config.yml
(http://www.dropwizard.io/0.9.2/docs/manual/core.html#environment-variables).
By making this non-strict we are able to provide the default log levels,
which then could be overridden by environment variables.

* The log levels are configured as environment variables with defaults.
This can be found in the environment variable settings in pay-scripts and
pay-ansible. The available log levels for eclipselink are at -
https://wiki.eclipse.org/EclipseLink/Examples/JPA/Logging