PP-582 Add logging filter
* Introducing a Logging filter that would log all incoming requests with
a prefix. Once the request is processed by the relevant resource it would
also log the end of the chain with a prefix.
* Any exception thrown will also be logged with a prefix, exception
message and the stack trace. A catch on this filter was opted against a
ExceptionMapper so we can get a nice trace of start to finish (with
erros)
* This gives us a nice trace of the request in the form of;
start->error(if any)->end