PP-447 Allow patching of email on charge
A user may wish to add an email to a charge after the charge is created.
This work adds a patch method to the charge/{chargeId} endpoint to allow this.
At the moment the only patch operation allowed is "replace" and the only
patchable field on the Charge entity is the email field.