PP-448 Send payment confirmation emails using GOV.UK notify
- This relies on the pay-notify java client jar which is not published
yet, so is built locally and loaded from the source folder as a maven dependenacy instead till we have it
published.

- Passes in the configuration from the environment variables to set up
the client to make a call to notify to send email confirmation message

- These emails are atm restricted to be internal domain emails only by
pay-notify

- Email notifications can be disabled using the environment variable
“NOTIFY_EMAIL_ENABLED”, set it to true to enable sending emails,
defaults to false.

- These emails are atm restricted to be internal domain emails only by
pay-notify

- allows you to pass  personalisation of the template that you choose.

- NotifyClient.jar was not getting into the all-in-one jar as it was
scoped system and hence getting ignored in the shaded jar. Fixed now.