PP-720 Return 2xx code when gateway cancel fails
Previously we returned a 400 when gateway cancel failed.
From the user/service point of view it is better to
return a success code, and then actually cancel the
charge at a later date. Even though the cancel has failed
with the gateway, the key thing s that the user can no
longer interact with the charge.