PP-1776 Refactor ClientFactoryTest to not spin up postgres
DropwizardAppWithPostgresRule spins up postgres in a docker container and runs
migrations. We don't need any of that for these tests.

Prior to this commit there was an elaborate use of paramterised tests in order
to pass the `enableProxy` flag through to the Dropwizard app rule's constructor.
This was based on the assumption that the only way to use the DropwizardAppRule
was to hook it into jUnit's lifecycle by using a `@Rule` annotation. However,
it turns out that it's quite straightforward to also use DropwizardTestSupport
explicitly by calling `.before()` and doing that we can simplify the injection
of the `enableProxy` flag.

Prior to this commit we were spinning up and tearing down postgres twice for
the two tests. With this change the tests should be about 10 seconds faster.