PP-1776 Set charge status to CAPTURE_APPROVED on error
The `CAPTURE_APPROVED` status represents a charge which the user has approved
for capture, but which we have not yet attempted to capture.

We are representing `CAPTURE_APPROVED` externally as an error in this commit to
preserve backwards compatibility.

Ultimately we intend `CAPTURE_APPROVED` to represent success externally. We'll
change the semantics of this once we've implemented the retry capability.