PP-1776 Allow capture approved charges to be captured
The `CardCaptureService` has a precondition guard on the status of charges
which are allowed to be captured. This is useful for preventing calls to the
(private) `/frontend/charges/{chargeId}/capture` endpoint from attempting to
capture a charge which should not be captured.

We overlooked this guard in the original work because the
`CardCaptureProcessTest` mocks out the `CardCaptureService`, and there is no
integrating test.