PP-1789 Including request id in each log entry
This change aims at improving logging for connector by solving 2 problems:

1. not all the logging entries had a requestId
2. even when the requestId was included in the log line, it was not clearly marked as the request id

We solved issue 1) by moving the inclusion of the requestId from the LoggingFilter to the logging configuration (in config/config.yaml) - this makes sure that every time Logger::info() (or the other levels) are called in the code, they are affected by the new format in the configuration. In fact, in the previous implementation the requestId was included only when the LoggingFilter::filter method was invoked.

We solved issue 2) by adding the string requestID= just before the request id.