PP-1897 Fixed a bug in the metric to track the size of the queue of charges ready for capture.
A Meter was used for this purpose, but a Meter is to measure the rate of events over time.
We rather need to just log the size of the queue overtime.
Among the Metrics available in Dropwizard, probably the Gauge is the more appropriate,
but that is not very efficient as that would need to calculate the size of the queue (which changes
very slowly) at regular interval. A CachedGauge is then a better option, but that complicates
things a bit and we still don't have control on when the metric collection is triggered.
That's why I have decided to use a Counter, which still makes sense, but we need to use it
a bit cleverly, as it just provide methods to increment or decrement the counter. Ideally,
to consider in the futre, is to create a new Metric (probably inheriting from Counter) which
would allow us to set a value directly with a set method.

For more information, you can see Dropwizard Metrics guide.