Revert "Merge pull request #373 from alphagov/PP-1775_reverse_commit_for_pull_request_368"
This reverts commit 3a7966fbda6c3e04d364acdeedeb6be5343d7f2e, reversing
changes made to 94f571f5fa4c4c9ad96243bfff867990b8b11a2e.

Reinstate what was previously reverted on master by pull request #373,
which is the change to start processing `CAPTURE` operation
asynchronously in the background. This code was temporarily reverted on
master so that in the meantime a feature flag for the new functionality
could be build (@see PP-1948).