PP-1989 Increase authorisation timeout to 50 seconds (from 20)
This controls how long we wait for the payment gateway to respond
when making an authorisation request.

Services were reporting a small number of AUTHORISATION ERRORs
(caused by GATEWAY_CONNECTION_TIMEOUT_ERRORs) with the previous 20
second limit.

The new limit of 50 seconds is chosen to be just below the 60
seconds used by the egress proxies.

with @heathd