PP-2042 Allow `CAPTURE_APPROVED_RETRY` to `CAPTURED`
We are seeing errors in logs of type
```
“message”:“2017-05-10T14:57:55+00:00 172.17.0.1 pay-connector[14620]: [2017-05-10 14:57:55.227]
[dw-17 - POST /v1/api/notifications/worldpay] \u001b[1;31mERROR\u001b[0;39m
\u001b[36mu.g.p.c.s.NotificationService\u001b[0;39m [requestID=xxxxxx] - Notification with
transaction id=xxxxxx failed updating charge status to: CAPTURED.
Error: Charge state transition [CAPTURE APPROVED RETRY] -> [CAPTURED] not allowed “
```

which indicates that in some cases we are receiving notifications that payments have been captured
that we still have in `CAPTURE_APPROVED_RETRY` status. We assume this is because
we are sending a capture request that times out from our point of view but succeeds with worldpay. See:
https://github.com/alphagov/pay-connector/blob/master/src/main/java/uk/gov/pay/connector/service/CardCaptureService.java#L122

This commit allows the transition from `CAPTURE_APPROVED_RETRY` to `CAPTURED` as this should be a valid
transition.