PP-2035 Extended the code so that we can handle different media types for each gateway.
This change was required because, whilst all our existing gateways (e.g. Worldpay)
requires 'application/xml' as the mediatype for requests, the new Epdq gateway
requires a different one (multipart/form-data).

On an implementation level, we removed mediatype as constructor argument for
GatewayClient (which has not got a per-gateway specific implementation), and moved
that information into the abstract method OrderRequestBuilder::getMediaType() which
has to be overridden by each per-gateway subclasses, e.g. EpdqOrderRequestBuilder.

with @alexbishop1