PP-2022 upgrading to notify client v3.x
The upgrade needed to be done in order to consistent with adminusers and to be compatible with the latest API from notify. The primary change is that they have moved in favour of a API Key style auth mechanism over the existing serviceId + secret approach.

Currently not happy with the Asynchronous implementation in UserNotificationService.java. This uses a fixed unmanaged threadpool + uses a Future<T> style to make it asynchronous . This could be improved to a more promise style approach (see adminusers notify integration style) and using a Dropwizard managed execution service.