PP-1500 Rework how we process incoming notifications
NotificationService now has four well-defined steps: parse, verify, evaluate,
update:
* Parse: parse payload into one or more notifications.
* Verify: check notification has necessary information and, in the case of
  ePDQ, check the SHA signature is correct.
* Evaluate: decide what to do with the notification (if it maps to a status we
  wish to apply to the payment). Unlike before, this happens after the payment
  is retrieved from the database so can take account of its current status.
* Update: modify the payment in the database.

StatusMapper is expanded beyond doing simple mappings from values in
notifications to a ChargeStatus or RefundStatus. While this is still supported,
it is now also possible to decide what to map to based on the value in the
notification and the current status of the payment. This is supported by new
StatusMapFromStatus objects to represent the two types of keys:
ProviderStatusOnly (key is just the value in the notification) and
ProviderStatusWithCurrentStatus (key is the value in the notification in
combination with the current status of the payment). When doing lookups, the
latter is always used. For this reason, a ProviderStatusOnly and a
ProviderStatusWithCurrentStatus are considered equal if they have the same
provider status (value in the notification).

This allows us to support ePDQ cancellation notifications, where the status
they provide us in the notification is "6" but we need to know the current
status of the payment in order to decide whether it should transition to
USER_CANCELLED, SYSTEM_CANCELLED or EXPIRED.

Everything related to resolving deferred statuses (and indeed the entire
concept of deferred status mappings) is removed (this was the initial attempt
to handle the ePDQ cancellation notifications).

MappedStatus, UnknownStatus and IgnoredStatus moved out of BaseStatusMapper
into top-level classes. Additionally, MappedStatus split into two:
MappedChargeStatus and MappedRefundStatus.

ExtendedNotification renamed to EvaluatedNotification, which is an interface
with two implementations: EvaluatedChargeStatusNotification and
EvaluatedRefundStatusNotification (no longer used to represent notifications
that do not map to a payment status modification i.e. notifications that are
ignored or unknown).

NotificationService no longer uses TransactionFlowProvider.

with @ckalista