PP-2032: WIP Spike connector sending metrics
This PR makes the DatabaseHealthChecker have another function.

It now also does a query to pg_stats_database and spits them out to
the metrics.

If it can't do this it puts a value of 0 in the stats_healthy metric.

It relies on your healthcheck being called frequently to update the healthy
stats of the database.