PP-2422 Send ePDQ refund requests and handle immediate responses
Sends RFD (partial refund) requests to ePDQ. Expects either status 81 (refund
pending) or 71 (deletion pending) responses (usually be 81 but can be 71 if the
capture has not yet been processed), which become REFUND SUBMITTED. Other
responses become REFUND ERROR.

Refund references are constructed by concatenating the PAYID (gateway
transaction ID) with a solidus followed by the PAYIDSUB (number that increments
with each operation) from the response. The same PAYID and PAYIDSUB will be
included in the subsequent notification, which will allow us to do the lookup by
reference.

Pseudo feature flag (hardcoded boolean set to true for tests and false for real)
used to stop refunds from working because we haven’t added support for handling
the subsequent notifications yet. This means that pressing the refund button in
selfservice or using publicapi will continue to display an error as it currently
does.

with @juanjoqmelian
with @alexbishop1