Refactor CardAuthoriseServiceTest
this is partly a 'comprehension refactoring'[1], but also an attempt to make
the tests more readable/easy to understand by moving towards a more consistent
given/when/then pattern for each test.

In my opinion, this structure makes it easier to understand a unit test by clearly separating out these elements:

1. setting up a stable and repeatable context within which the subject of the
   test will be exercised. This could involve setting up stubs or other test
   doubles, injecting these doubles or otherwise constructing objects which
   comprise the fixed and repeatable environment of the test (aka 'given')
2. explicitly invoking the subject of the test to trigger the tested behaviour (aka 'when')
3. inspecting return values and/or side effects of probing the test subject
   typically by making assertions or verifying mock interactions (aka 'then')

I have deliberately refactored helper methods which contained a combination of 'given' and 'when'.

For example the `authorisationResponse` method both configured the mock payment
provider to return a particular response ('given') and invoked the
`doAuthorise` method of `cardAuthorisationService` ('when'). I've now separated
all of these helper methods so that they only do mock setup (ie. purely
'given') and do not invoke the test subject.

Every test now contains an explicit invocation of the test subject `cardAuthorisationService.doAuthorise()`.

I've also renamed and refactored some of the fixture/mock setup methods.
Instead of `mockXXX` I've followed a pattern of naming `providerWillX`.

[1] https://martinfowler.com/articles/workflowsOfRefactoring/#comprehension