PP-2515 Mechanism to switch between fetching strategies
introduce strategies for fetching transactions:

- `TransactionSearchStrategy` fetches both charges and refunds
- `ChargeSearchStrategy` fetches charges only, using the original implementation

The strategy is selected according to the `REFUNDS_IN_TX_LIST` feature flag.

Testing:
- `TransactionDaoITest` exercises the query operations of the TransactionDao
- `ChargesApiResourceITest` exercises all layers from resource down, with `REFUNDS_IN_TX_LIST` disabled
- `TransactionsApiResourceITest` exercises all layers from resource down, with `REFUNDS_IN_TX_LIST` enabled

The strategy and switching via feature flag are temporary. We intend to
replace the `TransactionSearchStrategy` implementation after refactoring the
data model to store charges and refunds in the same database table.