PP-2627 Change the maximum number of capture retries from 24 to 48
If there is a connection timeout when sending a capture request to a payment
gateway, it’s possible for us to end up in a situation where the payment
gateway has processed the capture request but we think it has not. In this
case, we will retry every hour. We’ll stop if we get a successful response to
one of our requests or receive a notification telling us the payment has been
captured. If neither of these two things happen, after about 24 hours we’ll
give up and change the payment status to CAPTURE ERROR.

In a situation like the above (the payment gateway has processed the capture
request but we think it has not), Worldpay will respond to our capture retry
with a successful response because they treat capture requests as an idempotent
operation. ePDQ, however, appears to not work in the same way and will respond
with an error. This means that the only way we’ll break out of the retry cycle
is if we receive a capture notification.

Usually, ePDQ processes captures overnight. This means that we should receive a
capture notification during our 24 hour retry window. However, if a payment
happens very late in the day, it won’t get captured until the day after, which
means the notification could be received after we’ve given up retrying. By this
point, the payment status will be CAPTURE ERROR so we will ignore the
notification (we do not allow transitions from CAPTURE ERROR to CAPTURED).

Changing the maximum number of retries from 24 to 48 means we should still be
retrying when we receive the ePDQ capture notification.