Only run unit tests on PR and e2e on master
At the moment a making a connector PR into a release candidate
(ie. successfully deployed to test) involves the following tests:

- PR
  - unit tests (~10 mins)
  - e2e tests + accept tests (~ 20 mins)
- Master (eg. https://build.ci.pymnt.uk/job/pay-connector/job/master/)
  - unit tests (~10 mins)
  - e2e tests + accept tests + zap tests (~ 25 mins)

For  a total test time PR -> deploy of 65 mins
(Let's not start on the smoke tests)

In this commit I have changed this to the following
- PR
  - unit tests (~10 mins)
- Master (eg. https://build.ci.pymnt.uk/job/pay-connector/job/master/)
  - e2e tests + accept tests + zap tests (~ 25 mins)

For a total test time of 35 mins. \o/

The removal of unit tests on master build is risk free, as we  have github
ensure that only freshly rebased branches can be merged (so currently we
essantially run unit tests twice against *exactly* the same code).

The removal of PR end to end tests is slightly more controversial - it makes
it possible to commit potentially broken (or incompatible) code to master.
However, I believe developer discipline and intuition should guide whether
running this *before* merge is necessary or not. For small changes (which
we should be encouraging) this seems way too much largely irrelevant
testing. And if you do break master, it would be much quicker to get out
a fix. If you feel need for Jenkins to run e2e againts your branch before
merge, I have added a manually configurable option to be able to do that.