PP-2777 Create transactions table
- Mapped Transactions table to Transactions entity which links to the
PaymentRequestEntity.
- Made TransactionEntity reference PaymentRequestEntity. This creates a
bidirectional relationship which means the transaction can be inserted
with a single sql statement. Otherwise JPA will try to insert the
Transaction with a null in the payment_request_id column, then create
another statement to update the the payment_request_id.
- Moved chargeStatusUpdate to ChargeService from ChargeFrontendResource.
- Update Transactions when change status.
- Setting gatewayTransactionId on the chargeTransaction in the same
places as it is set on the charge.

with @georgeracu