PP-2777 Create transactions table
Check that a PaymentRequest has a charge transaction before modifying
it. This is needed until we back fill the transactions table with
anything from charges. Once that has been done this commit can be
reverted. Have done it this way as it is a relatively small change which
will be easy to revert. Thought about creating a new ChargeTransaction
when you call getChargeTransaction if it did not exist. However this
required a ChargeEntity to be passed to getChargeTransaction() and the
changes were becoming considerable. Want this to be easy to remove once
the transactions table has been back filled.

- Add a hasChargeTransaction() to PaymentRequestEntity.
- Check hasChargeTransaction() before making a change to the charge
transaction.