PP-2779 Create transaction_events table
Started to use the new transaction_events table. Whenever you update
the status of a TransactionEntity then a new TransactionEvent is added.
TransactionEventEntity is abstract and implemented by
ChargeTransactionEventEntity to allow a ChargeStatus to be set. We
will have a corresponding RefundTransactionEventEntity when we add
refunds to the transactions table which will have a RefundStatus on it.