PP-2778 Dual write refunds to transaction table
Start writing to the transactions table whenever we do a refund.
Have also started writing refund events to the
transaction_events table which had been broken out to a
different story. This is because this code is shared with
ChargeTransactions and was easier to leave in than disable.

- Created RefundTransactionEntity and
RefundTransactionEventEntity.
- Created RefundStatusUpdater to update status on
RefundTransactionEntity.
- Call method on RefundStatusUpdater whenever we we changing
the status on RefundEntity.
- Set refundReference on RefundTransactions once we have a response
from the gateway.