Suppress verbose logging during tests
For some of our tests, especially integration tests, there's quite a lot of
setup code which runs before dropwizard has booted. This means that
dropwizard's configuration system hasn't kicked in and hasn't read the logging
config from confg.yaml.

So if we drop in a logback-test.xml config into the class path it will get
picked up by logback's configuration engine and used as the standard
configuration for logging when tests are running.