PP-2783 Drop cards table charge_id column default, index and constraints
We want to stop writing to the charge_id column in the cards table (because
we’re going to start writing to the transaction_id column instead). However, it
has a unique index, NOT NULL constraint and — incorrectly — a default value
from a sequence. This means that if we stop inserting data into the charge_id
column, it will be filled with a meaningless default value from the sequence.
In addition to being completely wrong, the value might, by chance, conflict
with an existing charge_id value, violating the unique key.

This makes the following changes (in order):

- Drop the default value from charge_id column
- Drop the cards_charge_id_seq sequence used by the aforementioned default
- Drop the (unique) index on the charge_id column
- Drop the NOT NULL constraint on the charge_id column
- Drop the foreign key constraint that makes the charge_id column reference the
  id column in the charges table

For now, we will continue to write to charge_id column.

with @chrisgrimble