Remove state transitions tests
these tests are [tautological
tests](http://randycoulman.com/blog/2016/12/20/tautological-tests/). They are
testing that:

a) the state machine implementation matches the transitions written in the
test. In order to test this we must encode the full list of transitions in the
test case. At best this is 'double entry book keeping'. It makes it hard to
modify the state machine.

b) the transitions enforced by the state machine are actually enforced in
`ChargeEntity.setStatus()`. We don't really need to exhaustively test this, so
just add a 'happy' and 'sad' path test case. Ideally we should just test that
the statemachine is called by injecting a mock.