State machine implementation using immutable directed value graph
This data structure is a better fit with a state machine.

It allows us to do all of the various checks we would want to do with a state machine.

The 'value' part also allows us to store an arbitrary value for each 'edge' in
the graph (an edge corresponds to a state transition).

In the future we might store a richer data structure such as an object
representing an event.

Some light weight tests. We don't need to exhaustively test every state transition.

Removed the previous implementations