PP-3070 Use Windows-1252 rather than UTF-8 for ePDQ payloads
When constructing application/x-www-form-urlencoded payloads to send
to ePDQ, use Windows-1252 encoding instead of UTF-8.

It is our belief that ePDQ assumes incoming
application/x-www-form-urlencoded payloads are encoded in
Windows-1252 and decodes them as such. (ePDQ have not yet officially
confirmed this but our discussions and testing indicate this is the
case.)

With this change, we can successfully send payloads containing
characters like ’ U+2019 RIGHT SINGLE QUOTATION MARK (which encodes
to %E2%80%99 in UTF-8 but %92 in Windows-1252) in payloads. We know
that users with such characters in their names etc. have problems
making payments (the first hurdle that’s hit is the SHASIGN signature
verification fails: %E2%80%99 decodes to â€™ in Windows-1252, so
“John O’Connor” becomes “John Oâ€™Connor”, which has a different
SHA-512 digest).

This change makes no difference for the majority of payloads because
UTF-8 and Windows-1252 encode all printable ASCII characters in the
same way.

When generating the SHASIGN signature, we continue to take the UTF-8
bytes of the string, generate a SHA-512 digest of those bytes and
return the result as a hexadecimal string. (We use UTF-8 because
that’s what we always set in our ePDQ accounts at Configuration →
Technical information → Global security parameters → Hashing method,
alongside SHA-512.)