PP-2783 Add relationship between cards and transactions
Currently the cards table references the charge_id. As we will be
deleting the charges tables we want to reference the transaction_id of
the ChargeTransaction. The column has already been added and for new
entires in cards is being populated. This will back fill the entries in
cards that are missing the transaction_id with a Task which can be run
from the admin port of the app.