PP-2784 Add relationship between card_3ds and transactions
Currently the card_3ds table references the charge_id. As we will be
deleting the charges tables we want to reference the transaction_id of
the ChargeTransaction. This is the first step where we will add the new
column to the database. In the next release we will back fill and start
writing to the column.