PP-3239 Get ready to use new data structure in search
The new search is a bit slower than we would like. One of the biggest
reasons for this is joining payment_requests to transactions when we
have a large amount of data. This is the case if you just load
transactions for a given gateway_account_id (if you search by reference
also on payment_requests you are reducing the amount of rows by enough
for this not to be inefficient). Adding the gatewayAccountId to
TransactionEntities when we add the Transactions to the
PaymentRequestEntity. We probably do not want expose it in the Java, we
will just use it when searching transactions.

- Added the gatewayAccountId to TransactionEntity and set during
creation.
- Added back fill to set gatewayAccountId on TransactionEntities from
PaymentRequestEntity.