PP-3239 Rename TransactionDao and TransactionSearchStrategy
As we now have a table called Transactions that contains both charges
and refunds, we created a Java class called TransactionEntity. When we
went to create the DAO and search strategy for this entity we found
there was a TransactionDao and TransactionsSearchStrategy. This was
used by the old search which combined items from the Charges and
Refunds tables. Renamed these to OldTransactionDAO and
OldTransactionSearchStrategy.