Inline all resource path constants
The API paths form a primary part of the external API of the service.
Using constants to declare those paths makes it harder to read the
resource classes and understand what the actual endpoints are. I have
often found myself jumping back and forth between the API paths file and
the resource file to try to figure out which URL is used to trigger a
particular resource.

The only place apart from the resources where the paths were used is in
the tests. This effectively makes the checks tautologous -- in other
words you could change the path constant and the tests would still pass,
but in fact you would have broken dependent apps. This feels like a
disadvantage to me.

So, I think having paths inline in the resources (and explicitly stated
in the tests) makes things clearer, easier to read and simpler.